refactor: 기능 연결 및 파일 경로 수정
