+++ src/getRandomNumbers.js
@@ -1,9 +1,9 @@
 const MissionUtils = require('@woowacourse/mission-utils');
-const { BASEBALL_RANGE } = require('../constantValue');
+const { BASEBALL_RANGE } = require('./constantValue');
 
 function getRandomNumbers() {
     const Numbers = [];
-    while (Numbers.length < GENERATE_RANGE.MAX_SIZE) {
+    while (Numbers.length < BASEBALL_RANGE.MAX_SIZE) {
         const tmpNumber = MissionUtils.Random.pickNumberInRange(
             BASEBALL_RANGE.MINIMUM_VALUE,
             BASEBALL_RANGE.MAXIMUM_VALUE,
